+++ library/src/main/java/com/github/gfx/android/orma/adapter/DateAdapter.java
@@ -23,6 +23,7 @@ import java.util.Date;
  * Handles {@link Date} as an integer, or <strong>milliseconds</strong>.
  * Use {@link java.sql.Timestamp} if you want to use SQLite date/time functions.
  */
+@Deprecated
 public class DateAdapter extends AbstractTypeAdapter<Date> {
 
     @NonNull