+++ library/src/test/java/com/github/gfx/android/orma/test/model/ModelWithDeprecatedTypeAdapter.java
@@ -13,37 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package com.github.gfx.android.orma.test.model;
 
 import com.github.gfx.android.orma.annotation.Column;
 import com.github.gfx.android.orma.annotation.Table;
+import com.github.gfx.android.orma.test.toolbox.IntTuple2x;
 
-/**
- * @see ModelWithPrimitives_Schema
- *
- * @see ModelWithBoxTypes
- */
 @Table
-public class ModelWithPrimitives {
+public class ModelWithDeprecatedTypeAdapter {
 
     @Column
-    public boolean booleanValue;
-
-    @Column
-    public byte byteValue;
-
-    @Column
-    public short shortValue;
-
-    @Column
-    public int intValue;
-
-    @Column
-    public long longValue;
-
-    @Column
-    public float floatValue;
-
-    @Column
-    public double doubleValue;
+    public IntTuple2x intTuple2;
 }