+++ CHANGELOG.md
@@ -8,6 +8,23 @@ The versioning follows [Semantic Versioning](http://semver.org/):
 > * MINOR version when you add functionality in a backwards-compatible manner, and
 > * PATCH version when you make backwards-compatible bug fixes.
 
+## NEXT
+
+### New Features
+
+* Introduce Static type adapters by `@StaticTypeAdapter` (#131)
+  * The best-matched SQLite storage type is adapted
+    * e.g. `INTEGER` for `java.util.Date`, `BLOB` for `java.nio.ByteBuffer`
+  * Custom binary classes (e.g. `ByteBuffer`, `Bitmap`) are handled correctly
+    * v1.0.0 can't use binary objects except for `byte[]`
+
+## v1.1.3 - 2016/01/31
+
+### Bug Fixes
+
+- Use `com.tunnelvisionlabs:antlr4:4.5` instead of `org.antlr:antlr4:4.5.1`
+  because Android DataBinding 1.1 depends on com.tunnelvisionlabs's runtime which conflicts on org.antlr's runtime (#133)
+
 ## v1.1.2 - 2016/01/27
 
 ### Bug Fixes